+++ src/main/java/com/currency/exchange/dto/response/ExchangeResponseDto.java
@@ -2,6 +2,8 @@ package com.currency.exchange.dto.response;
 
 import com.currency.exchange.model.Currency;
 
-public record ExchangeResponseDto(CurrencyResponseDto baseCurrency, CurrencyResponseDto targetCurrency, double rate, double amount,
-                                  double convertedAmount) {
+import java.math.BigDecimal;
+
+public record ExchangeResponseDto(CurrencyResponseDto baseCurrency, CurrencyResponseDto targetCurrency, BigDecimal rate, BigDecimal amount,
+                                  BigDecimal convertedAmount) {
 }
